Introduce WarningToError class instead of string[]
Use default WarningToError property, can be changed it later.
